+++ web/client/utils/MapUtils.js
@@ -388,6 +388,12 @@ function getSimpleGeomType(geomType = "Point") {
     }
 }
 
+const getIdFromUri = (uri, regex = /data\/(\d+)/) => {
+    const decodedUri = decodeURIComponent(uri);
+    const findDataDigit = regex.exec(decodedUri);
+    return findDataDigit && findDataDigit.length && findDataDigit.length > 1 ? findDataDigit[1] : null;
+};
+
 module.exports = {
     EXTENT_TO_ZOOM_HOOK,
     RESOLUTIONS_HOOK,
@@ -417,5 +423,6 @@ module.exports = {
     saveMapConfiguration,
     isSimpleGeomType,
     getSimpleGeomType,
-    extractTileMatrixSetFromLayers
+    extractTileMatrixSetFromLayers,
+    getIdFromUri
 };
