+++ web/client/utils/__tests__/MapUtils-test.js
@@ -26,7 +26,8 @@ var {
     getBbox,
     getCurrentResolution,
     saveMapConfiguration,
-    extractTileMatrixSetFromLayers
+    extractTileMatrixSetFromLayers,
+    getIdFromUri
 } = require('../MapUtils');
 
 describe('Test the MapUtils', () => {
@@ -885,6 +886,14 @@ describe('Test the MapUtils', () => {
             }
         });
 
+    });
+    it('test getIdFromUri ', () => {
+        // /mapstore2/rest/geostore/data/578/raw?decode=datauri
+        expect(getIdFromUri('%2Fmapstore2%2Frest%2Fgeostore%2Fdata%2F578%2Fraw%3Fdecode%3Ddatauri')).toBe('578');
+        // rest/geostore/data/578/raw?id=1568321658464
+        expect(getIdFromUri('rest%2Fgeostore%2Fdata%2F578%2Fraw%3Fid%3D1568321658464')).toBe('578');
+        // rest/geostore/data/
+        expect(getIdFromUri('rest%2Fgeostore%2Fdata%2F')).toBe(null);
     });
 
 });
